working solution for relative addressing

+++ src/prelude.rs
@@ -3,4 +3,4 @@ pub use crate::endiannes::Endiannes;
 pub use crate::config::Config;
 // TODO add INSTR_LEN if we do that
 
-pub const BYTE_SIZE: u32 = 8;
+pub const BYTE_SIZE: u64 = 8;
